update makeLoader call to new sig; bump versions
