+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "databyss-org-workspace",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "description": "Expressive database management tools for writers and researchers",
   "main": "index.js",
   "repository": {
