+++ packages/databyss-api/package.json
@@ -1,13 +1,13 @@
 {
   "name": "@databyss-org/api",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "description": "api",
   "main": "server.js",
   "author": "databyss-org",
   "license": "MIT",
   "peerDependencies": {
     "@bugsnag/js": "^6.3.2",
-    "@databyss-org/services": "1.1.71"
+    "@databyss-org/services": "1.1.72"
   },
   "dependencies": {
     "@babel/core": "^7.5.0",
