+++ packages/databyss-documentation/package.json
@@ -1,6 +1,6 @@
 {
   "name": "databyss-documentation",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "description": "documentation",
   "license": "MIT",
   "scripts": {
