+++ packages/databyss-notes-native/package.json
@@ -2,7 +2,7 @@
   "name": "@databyss-org/notes-native",
   "author": "databyss-org",
   "license": "MIT",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "private": true,
   "repository": {
     "type": "git",
@@ -18,7 +18,7 @@
   },
   "peerDependencies": {
     "react": "16.9.0",
-    "@databyss-org/ui": "1.1.71"
+    "@databyss-org/ui": "1.1.72"
   },
   "publishConfig": {
     "access": "public"
