+++ packages/databyss-notes/package.json
@@ -1,12 +1,12 @@
 {
   "name": "@databyss-org/notes",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "private": true,
   "peerDependencies": {
     "react": "16.9.0",
     "react-dom": "16.9.0",
     "@reach/router": "^1.3.3",
-    "@databyss-org/ui": "1.1.71",
-    "@databyss-org/services": "1.1.71"
+    "@databyss-org/ui": "1.1.72",
+    "@databyss-org/services": "1.1.72"
   }
 }
