+++ packages/databyss-services/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@databyss-org/services",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "repository": {
     "type": "git",
     "url": "https://github.com/databyss-org/databyss.git"
@@ -12,8 +12,8 @@
   "license": "MIT",
   "private": true,
   "peerDependencies": {
-    "@databyss-org/ui": "1.1.71",
-    "@databyss-org/services": "1.1.71",
+    "@databyss-org/ui": "1.1.72",
+    "@databyss-org/services": "1.1.72",
     "@bugsnag/js": "^6.3.2",
     "react": "16.9.0",
     "react-dom": "16.9.0"
