+++ packages/databyss-stories-native/package.json
@@ -2,7 +2,7 @@
   "name": "@databyss-org/stories-native",
   "author": "databyss-org",
   "license": "MIT",
-  "version": "1.1.71",
+  "version": "1.1.72",
   "private": true,
   "repository": {
     "type": "git",
@@ -22,7 +22,7 @@
     "react-native-svg-transformer": "^0.13.0"
   },
   "peerDependencies": {
-    "@databyss-org/ui": "1.1.71",
+    "@databyss-org/ui": "1.1.72",
     "react": "16.9.0"
   },
   "publishConfig": {
