+++ packages/databyss-ui/stories/Services/sources.stories.js
@@ -4,8 +4,9 @@ import { View, Button, RichTextInput } from '@databyss-org/ui/primitives'
 
 import SourceProvider, {
   useSourceContext,
-  withSource,
 } from '@databyss-org/services/sources/SourceProvider'
+
+import { withSource } from '@databyss-org/ui/components/Loaders'
 
 import reducer, { initialState } from '@databyss-org/services/sources/reducer'
 import { ViewportDecorator } from '../decorators'
